Revert "Exported local var def information from type analysis."
This reverts commit d313cb25e739abdb27adfed5796977db0ae694e8.
